Update README.md
While working at the Internship Studio, I created an e-commerce website for clothing, electronics, and other products. Working as a front-end and back-end developer, I used HTML, CSS, and JavaScript for the front-end, and Node.js and MongoDB for the back-end. The project runs from March 28 to May 2, 2022. By the end of the project, we had successfully enhanced the user experience and functionality, demonstrated my expertise in web development, and allowed me to contribute to a team the purpose of the well.